refactor: moved map components to pages/map folder
