+++ sites/geohub/src/components/pages/map/plugins/LayerVisibilitySwitcher.svelte
